refactor : dto -> uiModel
