+++ app/src/main/java/woowacourse/movie/booking/BookingPresenter.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.booking
 import android.os.Bundle
 import woowacourse.movie.domain.MovieScheduleGenerator
 import woowacourse.movie.domain.TicketCount
-import woowacourse.movie.dto.MovieInfo
+import woowacourse.movie.uiModel.MovieInfo
 
 class BookingPresenter : BookingContract.Presenter {
     override fun onCreateView(
