+++ app/src/main/java/woowacourse/movie/uiModel/Ticket.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.dto
+package woowacourse.movie.uiModel
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
